Added the DataCleaner and StringCleaner
The string cleaner removes punctuation, capitalization, and attempts to remove pluralization without butchering the word too much. The Data cleaner cleans the data as described above, as well as changes all multi-word ingredients to a single word, based on which word in the multi-word ingredient is used most in the rest of the data set. I recommend testing it by printing a before and after cleaning, on the first or second recipe for example.